+++ tests/tests/wgpu-gpu/device.rs
@@ -419,7 +419,7 @@ static DEVICE_DESTROY_THEN_MORE: GpuTestConfiguration = GpuTestConfiguration::ne
                             ty: wgpu::BufferBindingType::Uniform,
                             has_dynamic_offset: false,
                             min_binding_size: std::num::NonZeroU64::new(
-                                (ctx.device.limits().max_uniform_buffer_binding_size * 2) as u64,
+                                ctx.device.limits().max_uniform_buffer_binding_size * 2,
                             ),
                         },
                         count: None,